requirements.txt
attrs==19.3.0
backcall==0.1.0
backports-datetime-fromisoformat==1.0.0
bleach==3.1.0
bokeh==2.1.1
boto==2.49.0
boto3==1.13.26
botocore==1.16.26
bounded-pool-executor==0.0.3
certifi==2020.6.20
chardet==3.0.4
Click==7.0
cloudpickle==1.5.0
colorama==0.3.9
cvxopt==1.2.5
cycler==0.10.0
Cython==0.29.14
dask==2.20.0
deap==1.3.1
decorator==4.4.1
defusedxml==0.6.0
dill==0.3.2
dispy==4.12.2
distributed==2.20.0
docutils==0.15.2
entrypoints==0.3
et-xmlfile==1.0.1
fsspec==0.7.4
future==0.18.2
gensim==3.8.3
hdbscan==0.8.23
HeapDict==1.0.1
idna==2.9
imageio==2.6.1
importlib-metadata==1.3.0
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jdcal==1.4.1
jedi==0.15.2
Jinja2==2.10.3
jmespath==0.10.0
joblib==0.14.0
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
jupyterlab==2.1.5
jupyterlab-server==1.1.5
karateclub==1.0.6
kiwisolver==1.2.0
locket==0.2.0
MarkupSafe==1.1.1
matplotlib
mistune==0.8.4
mmh3==2.5.1
more-itertools==8.0.2
msgpack==0.6.2
multiprocess==0.70.10
mysql-connector-python==8.0.18
nbconvert==5.6.1
nbformat==5.0.3
netifaces==0.10.9
networkx==2.4
nose==1.3.7
notebook==6.0.2
numpy==1.19.0
olefile==0.46
opencv-python==4.1.2.30
openpyxl==3.0.0
packaging==20.4
pandas
pandocfilters==1.4.2
parso==0.5.2
partd==1.1.0
pathos==0.2.6
pexpect==4.7.0
pickleshare==0.7.5
Pillow
pox==0.2.8
ppft==1.6.6.2
pqdm==0.1.0
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.9.1
psutil==5.6.7
ptyprocess==0.6.0
pycos==4.8.15
pydot==1.4.1
Pygments==2.5.2
PyGSP==0.5.1
Pympler==0.8
pyparsing==2.4.7
pyrsistent==0.15.7
python-dateutil==2.8.1
python-louvain==0.14
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.2
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.23.0
s3transfer==0.3.3
scikit-image==0.16.2
scikit-learn==0.23.1
scipy
seaborn==0.10.1
semantic-version==2.8.5
Send2Trash==1.5.0
setuptools-rust==0.10.6
simplejson==3.17.0
sip
six==1.15.0
sklearn==0.0
smart-open==2.0.0
sortedcontainers==2.1.0
style==1.1.0
tblib==1.6.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.1
toolz==0.10.0
tornado==6.0.4
tqdm==4.43.0
traitlets==4.3.3
typing-extensions==3.7.4.2
update==0.0.1
urllib3==1.25.3
wcwidth==0.1.8
webcolors==1.11.1
webencodings==0.5.1
widgetsnbextension==3.5.1
xlrd==1.2.0
zict==1.0.0
zipp==0.6.0
